+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project uses [semantic versioning](https://semver.org/)
 ## 0.3.0dev
 ### Added
 - New `device info` command to get device information
+- New `lte` group with cellular network controls
 
 ### Changed
 - Project structure enhanced to be more modular